python_controller/requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
websocket-client==1.6.1